coprocessor/led/pyproject.toml
[project]
name = "led"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "adafruit-circuitpython-neopixel>=6.4.1",
    "pyntcore>=2026.2.1",
    "RPi.GPIO>=0.7.1",
    "rpi-ws281x>=5.0.0"
]
